Converted BigDecimal to Float as excessive storage and precision are unneeded for level.

+++ src/main/java/com/nexmo/client/voice/ncco/TalkNcco.java
@@ -25,8 +25,6 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import java.math.BigDecimal;
-
 @JsonInclude(value = JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class TalkNcco implements Ncco {
@@ -35,7 +33,7 @@ public class TalkNcco implements Ncco {
     private String text;
     private Boolean bargeIn = null;
     private Integer loop = null;
-    private BigDecimal level = null;
+    private Float level = null;
     private String voiceName = null;
 
     public TalkNcco(@JsonProperty("text") String text) {
@@ -66,11 +64,11 @@ public class TalkNcco implements Ncco {
         this.loop = loop;
     }
 
-    public void setLevel(BigDecimal level) {
+    public void setLevel(Float level) {
         this.level = level;
     }
 
-    public BigDecimal getLevel() {
+    public Float getLevel() {
         return level;
     }
